Conflicts:
	src/server/game/AI/CoreAI/TotemAI.cpp
	src/server/game/Entities/Player/Player.cpp
	src/server/game/Entities/Unit/Unit.cpp
	src/server/game/Globals/ObjectMgr.cpp
	src/server/game/Guilds/Guild.cpp
	src/server/game/Guilds/Guild.h
	src/server/game/Movement/Spline/MoveSplineInit.cpp
	src/server/game/Movement/Spline/MovementPacketBuilder.cpp
	src/server/game/Movement/Spline/MovementPacketBuilder.h
	src/server/game/Spells/SpellMgr.cpp
	src/server/scripts/EasternKingdoms/BlackrockMountain/BlackwingLair/boss_nefarian.cpp
	src/server/scripts/EasternKingdoms/ZulAman/boss_akilzon.cpp
	src/server/scripts/EasternKingdoms/zone_silverpine_forest.cpp
	src/server/scripts/Kalimdor/zone_mulgore.cpp
	src/server/scripts/Kalimdor/zone_stonetalon_mountains.cpp
	src/server/shared/Database/Implementation/CharacterDatabase.cpp

+++ src/server/shared/Logging/Appender.h
@@ -75,7 +75,7 @@ struct LogMessage
     ///@ Returns size of the log message content in bytes
     uint32 Size() const
     {
-        return prefix.size() + text.size();
+        return static_cast<uint32>(prefix.size() + text.size());
     }
 };
